# schemas/generated/orchestration-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tasker Web API",
    "description": "REST API for the Tasker workflow orchestration system",
    "contact": {
      "name": "Tasker Systems",
      "email": "support@tasker-systems.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "API version 1"
    }
  ],
  "paths": {
...
    "/v1/dlq": {
      "get": {
        "tags": [
          "dlq"
        ],
        "summary": "List DLQ entries: GET /v1/dlq",
        "description": "Returns a list of DLQ entries with optional filtering by resolution status.\nSupports pagination via limit/offset parameters.\n\n# Query Parameters\n\n- `resolution_status` - Filter by resolution status (optional)\n- `limit` - Maximum entries to return (default: 50)\n- `offset` - Pagination offset (default: 0)\n\n# Example\n\n```text\nGET /v1/dlq?resolution_status=pending&limit=20\n```\n\n**Required Permission:** `dlq:read`",
        "operationId": "list_dlq_entries",
        "parameters": [
          {
            "name": "resolution_status",
            "in": "query",
            "description": "Filter by resolution status (optional)",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/DlqResolutionStatus"
                }
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of entries to return (default: 50)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of DLQ entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DlqEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "dlq:read"
      }
    },
    "/v1/dlq/entry/{dlq_entry_uuid}": {
      "patch": {
        "tags": [
          "dlq"
        ],
        "summary": "Update DLQ investigation status: PATCH /v1/dlq/entry/{dlq_entry_uuid}",
        "description": "Updates the investigation status and notes for a DLQ entry.\nThis endpoint tracks the INVESTIGATION workflow, not task resolution.\n\n**Note**: Task resolution happens via step APIs:\n`PATCH /v1/tasks/{uuid}/workflow_steps/{step_uuid}`\n\n# Path Parameters\n\n- `dlq_entry_uuid` - UUID of the DLQ entry\n\n# Request Body\n\n```json\n{\n  \"resolution_status\": \"manually_resolved\",\n  \"resolution_notes\": \"Fixed blocked step by recreating upstream dependency\",\n  \"resolved_by\": \"operator@example.com\"\n}\n```\n\n**Required Permission:** `dlq:update`",
        "operationId": "update_dlq_investigation",
        "parameters": [
          {
            "name": "dlq_entry_uuid",
            "in": "path",
            "description": "DLQ entry UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestigationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Investigation updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateInvestigationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "DLQ entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "dlq:update"
      }
    },
    "/v1/dlq/investigation-queue": {
      "get": {
        "tags": [
          "dlq"
        ],
        "summary": "Get DLQ investigation queue: GET /v1/dlq/investigation-queue",
        "description": "Returns a prioritized queue of pending DLQ entries for operator triage.\nEntries are ordered by priority score (higher = more urgent) combining\nbase reason priority with age factor.\n\n# Query Parameters\n\n- `limit` - Maximum entries to return (default: 100)\n\n# Response\n\nReturns investigation queue entries including:\n- DLQ entry and task UUIDs\n- Original state and DLQ reason\n- Task metadata (namespace, name, current state)\n- Time in DLQ (minutes)\n- Priority score for triage ordering\n\n# Example\n\n```text\nGET /v1/dlq/investigation-queue?limit=50\n```\n\n**Required Permission:** `dlq:read`",
        "operationId": "get_investigation_queue",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum entries to return (default: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prioritized investigation queue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DlqInvestigationQueueEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "dlq:read"
      }
    },
    "/v1/dlq/staleness": {
      "get": {
        "tags": [
          "dlq"
        ],
        "summary": "Get task staleness monitoring: GET /v1/dlq/staleness",
        "description": "Returns real-time staleness monitoring for active tasks in waiting states.\nProvides proactive visibility into tasks approaching staleness thresholds,\nenabling prevention of DLQ entries through early intervention.\n\n# Query Parameters\n\n- `limit` - Maximum tasks to return (default: 100)\n\n# Response\n\nReturns staleness monitoring entries with:\n- Task UUID and metadata (namespace, name)\n- Current state and time in state\n- Staleness threshold and health status classification\n- Health status: `healthy` (< 80%), `warning` (80-99%), `stale` (≥ 100%)\n\nResults are ordered by health status (stale first) then time in state.\n\n# Use Cases\n\n- **Alerting**: Monitor tasks at 80%+ threshold (warning status)\n- **Prevention**: Investigate tasks before they hit DLQ\n- **Capacity Planning**: Identify systemic staleness patterns\n\n# Example\n\n```text\nGET /v1/dlq/staleness?limit=50\n```\n\n**Required Permission:** `dlq:read`",
        "operationId": "get_staleness_monitoring",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum tasks to return (default: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task staleness monitoring data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StalenessMonitoring"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "dlq:read"
      }
    },
    "/v1/dlq/stats": {
      "get": {
        "tags": [
          "dlq"
        ],
        "summary": "Get DLQ statistics: GET /v1/dlq/stats",
        "description": "Returns aggregated statistics for DLQ entries grouped by reason.\nUseful for identifying systemic issues and patterns.\n\n# Response\n\nReturns statistics including:\n- Total entries per reason\n- Count by resolution status (pending, resolved, failed)\n- Oldest and newest entry timestamps\n\n# Example\n\n```text\nGET /v1/dlq/stats\n```\n\n**Required Permission:** `dlq:stats`",
        "operationId": "get_dlq_stats",
        "responses": {
          "200": {
            "description": "DLQ statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DlqStats"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key_auth": []
          }
        ],
        "x-required-permission": "dlq:stats"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DlqResolutionStatus": {
        "type": "string",
        "description": "DLQ Resolution Status\n\nTracks the lifecycle of a DLQ investigation (NOT task state).\nThis enum maps to the PostgreSQL `dlq_resolution_status` enum type.\n\nState Machine:\n- `pending` → `manually_resolved` (operator fixed problem via step APIs)\n- `pending` → `permanently_failed` (unfixable issue, task stays in Error state)\n- `pending` → `cancelled` (investigation no longer needed)\n\nA task can have multiple DLQ entries over time (investigation history),\nbut only one pending entry at a time (enforced by unique index).",
        "enum": [
          "Pending",
          "ManuallyResolved",
          "PermanentlyFailed",
          "Cancelled"
        ]
      },
...
      "UpdateInvestigationRequest": {
        "type": "object",
        "description": "Request body for updating DLQ investigation status",
        "properties": {
          "metadata": {
            "description": "Additional metadata (optional)"
          },
          "resolution_notes": {
            "type": [
              "string",
              "null"
            ],
            "description": "Investigation notes (optional)"
          },
          "resolution_status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DlqResolutionStatus",
                "description": "New resolution status (optional)"
              }
            ]
          },
          "resolved_by": {
            "type": [
              "string",
              "null"
            ],
            "description": "Who resolved the investigation (optional)"
          }
        }
      },
      "UpdateInvestigationResponse": {
        "type": "object",
        "description": "Response for update investigation endpoint",
        "required": [
          "success",
          "message",
          "dlq_entry_uuid"
        ],
        "properties": {
          "dlq_entry_uuid": {
            "type": "string",
            "format": "uuid"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key_auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}